// src/types.ts
import * as crypto from 'crypto';

export type HashAlgorithm = 'sha1' | 'sha256' | 'sha512';

export type CryptoCipher =
	| 'aes128'
	| 'aes192'
	| 'aes256'
	| 'aria128'
	| 'aria192'
	| 'aria256'
	| 'blowfish'
	| 'camellia128'
	| 'camellia192'
	| 'camellia256'
	| 'cast'
	| 'des'
	| 'des3'
	| 'idea'
	| 'rc2'
	| 'seed'
	| 'sm4';

export type CipherKeySize = 64 | 96 | 128 | 192 | 256 | 448;
export type CipherMode = 'cbc' | 'cfb' | 'ctr' | 'gcm' | 'ofb';

export type CryptAlgorithms = 'rsa' | 'dsa' | 'ec' | 'ed25519' | 'ed448' | 'x25519';
export type KeySize = 128 | 512 | 1024 | 2048 | 4096;

export type CryptoECCurve =
	| 'SM2'
	| 'brainpoolP160r1'
	| 'brainpoolP160t1'
	| 'brainpoolP192r1'
	| 'brainpoolP192t1'
	| 'brainpoolP224r1'
	| 'brainpoolP224t1'
	| 'brainpoolP256r1'
	| 'brainpoolP256t1'
	| 'brainpoolP320r1'
	| 'brainpoolP320t1'
	| 'brainpoolP384r1'
	| 'brainpoolP384t1'
	| 'brainpoolP512r1'
	| 'brainpoolP512t1'
	| 'c2pnb163v1'
	| 'c2pnb163v2'
	| 'c2pnb163v3'
	| 'c2pnb176v1'
	| 'c2pnb208w1'
	| 'c2pnb272w1'
	| 'c2pnb304w1'
	| 'c2pnb368w1'
	| 'c2tnb191v1'
	| 'c2tnb191v2'
	| 'c2tnb191v3'
	| 'c2tnb239v1'
	| 'c2tnb239v2'
	| 'c2tnb239v3'
	| 'c2tnb359v1'
	| 'c2tnb431r1'
	| 'prime192v1'
	| 'prime192v2'
	| 'prime192v3'
	| 'prime239v1'
	| 'prime239v2'
	| 'prime239v3'
	| 'prime256v1'
	| 'secp112r1'
	| 'secp112r2'
	| 'secp128r1'
	| 'secp128r2'
	| 'secp160k1'
	| 'secp160r1'
	| 'secp160r2'
	| 'secp192k1'
	| 'secp224k1'
	| 'secp224r1'
	| 'secp256k1'
	| 'secp384r1'
	| 'secp521r1'
	| 'sect113r1'
	| 'sect113r2'
	| 'sect131r1'
	| 'sect131r2'
	| 'sect163k1'
	| 'sect163r1'
	| 'sect163r2'
	| 'sect193r1'
	| 'sect193r2'
	| 'sect233k1'
	| 'sect233r1'
	| 'sect239k1'
	| 'sect283k1'
	| 'sect283r1'
	| 'sect409k1'
	| 'sect409r1'
	| 'sect571k1'
	| 'sect571r1'
	| 'wap-wsg-idm-ecid-wtls1'
	| 'wap-wsg-idm-ecid-wtls10'
	| 'wap-wsg-idm-ecid-wtls11'
	| 'wap-wsg-idm-ecid-wtls12'
	| 'wap-wsg-idm-ecid-wtls3'
	| 'wap-wsg-idm-ecid-wtls4'
	| 'wap-wsg-idm-ecid-wtls5'
	| 'wap-wsg-idm-ecid-wtls6'
	| 'wap-wsg-idm-ecid-wtls7'
	| 'wap-wsg-idm-ecid-wtls8'
	| 'wap-wsg-idm-ecid-wtls9';

export type Signatures =
	| 'RSA-MD4'
	| 'RSA-MD5'
	| 'RSA-MDC2'
	| 'RSA-RIPEMD160'
	| 'RSA-SHA1'
	| 'RSA-SHA1-2'
	| 'RSA-SHA224'
	| 'RSA-SHA256'
	| 'RSA-SHA3-224'
	| 'RSA-SHA3-256'
	| 'RSA-SHA3-384'
	| 'RSA-SHA3-512'
	| 'RSA-SHA384'
	| 'RSA-SHA512';